+++ packages/tailwindcss/src/css-functions.ts
@@ -13,11 +13,14 @@ export function substituteFunctions(ast: AstNode[], pluginApi: PluginAPI) {
       return
     }
 
-    // Find @media rules
+    // Find at-rules rules
     if (node.kind === 'rule') {
       if (
         node.selector[0] === '@' &&
-        node.selector.startsWith('@media ') &&
+        (node.selector.startsWith('@media ') ||
+          node.selector.startsWith('@custom-media ') ||
+          node.selector.startsWith('@container ') ||
+          node.selector.startsWith('@supports ')) &&
         node.selector.includes(THEME_FUNCTION_INVOCATION)
       ) {
         node.selector = substituteFunctionsInValue(node.selector, pluginApi)